fix chart view and add optional db toggle

+++ server/routes/tickerRoutes.ts
@@ -8,7 +8,7 @@ router.all('/get/tickers', async (req, res) => {
     let stock = req.app.get('stockTickers');
     let crypto = req.app.get('cryptoTickers');
 
-    if (!stock && !crypto) {
+    if (!stock && !crypto && process.env.DB) {
       console.log('Did not find any set tickers in express, querying mongodb');
       stock = (await Stocks.find()).map((ticker) => ticker.stock).toString();
       crypto = (await Cryptos.find()).map((ticker) => ticker.crypto).toString();
@@ -31,23 +31,27 @@ router.all('/set/tickers', async (req, res) => {
     let resString = '';
 
     if (cryptoTickers != undefined) {
-      await Cryptos.deleteMany({});
-      req.app.set('cryptoTickers', cryptoTickers);
-      const tickerArr = cryptoTickers.split(',');
-      for (const ticker of tickerArr) {
-        console.log(ticker);
-        await Cryptos.create({ crypto: ticker });
+      if (process.env.DB) {
+        await Cryptos.deleteMany({});
+        const tickerArr = cryptoTickers.split(',');
+        for (const ticker of tickerArr) {
+          console.log(ticker);
+          await Cryptos.create({ crypto: ticker });
+        }
       }
+      req.app.set('cryptoTickers', cryptoTickers);
       resString += `updated crypto tickers to be ${cryptoTickers}`;
     }
 
     if (stockTickers != undefined) {
-      await Stocks.deleteMany({});
-      req.app.set('stockTickers', stockTickers);
-      const tickerArr = stockTickers.split(',');
-      for (const ticker of tickerArr) {
-        await Stocks.create({ stock: ticker });
+      if (process.env.DB) {
+        await Stocks.deleteMany({});
+        const tickerArr = stockTickers.split(',');
+        for (const ticker of tickerArr) {
+          await Stocks.create({ stock: ticker });
+        }
       }
+      req.app.set('stockTickers', stockTickers);
       resString += `updated stock tickers to be ${stockTickers}`;
     }
